refactor: P3-T9 - 抽出CalibrationFlowViewModel
